io-gen/requirements.txt
pytest==8.3.4
pytest-cov
pyyaml

